Network.AWS.Polly.DeleteLexicon

Description

Deletes the specified pronunciation lexicon stored in an AWS Region. A lexicon which has been deleted is not available for speech synthesis, nor is it possible to retrieve it using either the GetLexicon or ListLexicon APIs.

For more information, see Managing Lexicons .

Creating a Request

deleteLexicon

Arguments

:: Text
dlName

-> DeleteLexicon

Creates a value of DeleteLexicon with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlName - The name of the lexicon to delete. Must be an existing lexicon in the region.

data DeleteLexicon

See: deleteLexicon smart constructor.

Request Lenses

dlName :: Lens' DeleteLexicon Text

The name of the lexicon to delete. Must be an existing lexicon in the region.

Destructuring the Response

deleteLexiconResponse

Arguments

:: Int
dlrsResponseStatus

-> DeleteLexiconResponse

Creates a value of DeleteLexiconResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlrsResponseStatus - -- | The response status code.

data DeleteLexiconResponse

See: deleteLexiconResponse smart constructor.

Response Lenses

dlrsResponseStatus :: Lens' DeleteLexiconResponse Int

- | The response status code.