Network.AWS.Polly.ListLexicons

Description

Returns a list of pronunciation lexicons stored in an AWS Region. For more information, see Managing Lexicons .

Creating a Request

listLexicons :: ListLexicons

Creates a value of ListLexicons with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

llNextToken - An opaque pagination token returned from previous ListLexicons operation. If present, indicates where to continue the list of lexicons.

data ListLexicons

See: listLexicons smart constructor.

Request Lenses

llNextToken :: Lens' ListLexicons (Maybe Text)

An opaque pagination token returned from previous ListLexicons operation. If present, indicates where to continue the list of lexicons.

Destructuring the Response

listLexiconsResponse

Arguments

:: Int
llrsResponseStatus

-> ListLexiconsResponse

Creates a value of ListLexiconsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

llrsLexicons - A list of lexicon names and attributes.

llrsNextToken - The pagination token to use in the next request to continue the listing of lexicons. NextToken is returned only if the response is truncated.

llrsResponseStatus - -- | The response status code.

data ListLexiconsResponse

See: listLexiconsResponse smart constructor.

Response Lenses

llrsLexicons :: Lens' ListLexiconsResponse [LexiconDescription]

A list of lexicon names and attributes.

llrsNextToken :: Lens' ListLexiconsResponse (Maybe Text)

The pagination token to use in the next request to continue the listing of lexicons. NextToken is returned only if the response is truncated.

llrsResponseStatus :: Lens' ListLexiconsResponse Int

- | The response status code.