Network.AWS.Polly.PutLexicon

Description

Stores a pronunciation lexicon in an AWS Region. If a lexicon with the same name already exists in the region, it is overwritten by the new lexicon. Lexicon operations have eventual consistency, therefore, it might take some time before the lexicon is available to the SynthesizeSpeech operation.

For more information, see Managing Lexicons .

Creating a Request

putLexicon

Arguments

:: Text
plName

-> Text
plContent

-> PutLexicon

Creates a value of PutLexicon with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

plName - Name of the lexicon. The name must follow the regular express format [0-9A-Za-z]{1,20}. That is, the name is a case-sensitive alphanumeric string up to 20 characters long.

plContent - Content of the PLS lexicon as string data.

data PutLexicon

See: putLexicon smart constructor.

Request Lenses

plName :: Lens' PutLexicon Text

Name of the lexicon. The name must follow the regular express format [0-9A-Za-z]{1,20}. That is, the name is a case-sensitive alphanumeric string up to 20 characters long.

plContent :: Lens' PutLexicon Text

Content of the PLS lexicon as string data.

Destructuring the Response

putLexiconResponse

Arguments

:: Int
plrsResponseStatus

-> PutLexiconResponse

Creates a value of PutLexiconResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

plrsResponseStatus - -- | The response status code.

data PutLexiconResponse

See: putLexiconResponse smart constructor.

Response Lenses

plrsResponseStatus :: Lens' PutLexiconResponse Int

- | The response status code.