Network.AWS.Polly.Types

Description

Service Configuration

polly :: Service

API version 2016-06-10 of the Amazon Polly SDK configuration.

Errors

_UnsupportedPlsLanguageException :: AsError a => Getting (First ServiceError) a ServiceError

The language specified in the lexicon is unsupported. For a list of supported languages, see Lexicon Attributes .

_InvalidSsmlException :: AsError a => Getting (First ServiceError) a ServiceError

The SSML you provided is invalid. Verify the SSML syntax, spelling of tags and values, and then try again.

_InvalidSampleRateException :: AsError a => Getting (First ServiceError) a ServiceError

The specified sample rate is not valid.

_MaxLexiconsNumberExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of lexicons would be exceeded by this operation.

_TextLengthExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The value of the Text parameter is longer than the accepted limits. The limit for input text is a maximum of 3000 characters total, of which no more than 1500 can be billed characters. SSML tags are not counted as billed characters.

_MaxLexemeLengthExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum size of the lexeme would be exceeded by this operation.

_InvalidLexiconException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Polly can't find the specified lexicon. Verify that the lexicon's name is spelled correctly, and then try again.

_ServiceFailureException :: AsError a => Getting (First ServiceError) a ServiceError

An unknown condition has caused a service failure.

_UnsupportedPlsAlphabetException :: AsError a => Getting (First ServiceError) a ServiceError

The alphabet specified by the lexicon is not a supported alphabet. Valid values are x-sampa and ipa .

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The NextToken is invalid. Verify that it's spelled correctly, and then try again.

_MarksNotSupportedForFormatException :: AsError a => Getting (First ServiceError) a ServiceError

Speech marks are not supported for the OutputFormat selected. Speech marks are only available for content in json format.

_SsmlMarksNotSupportedForTextTypeException :: AsError a => Getting (First ServiceError) a ServiceError

SSML speech marks are not supported for plain text-type input.

_LexiconSizeExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum size of the specified lexicon would be exceeded by this operation.

_LexiconNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Polly can't find the specified lexicon. This could be caused by a lexicon that is missing, its name is misspelled or specifying a lexicon that is in a different region.

Verify that the lexicon exists, is in the region (see ListLexicons ) and that you spelled its name is spelled correctly. Then try again.

Gender

data Gender

Constructors

Female

Male

LanguageCode

data LanguageCode

Constructors

CyGb

DaDk

DeDe

EnAu

EnGb

EnGbWls

EnIn

EnUs

EsEs

EsUs

FrCa

FrFr

IsIs

ItIt

JaJp

NbNo

NlNl

PlPl

PtBr

PtPt

RoRo

RuRu

SvSe

TrTr

OutputFormat

data OutputFormat

Constructors

JSON

MP3

OggVorbis

Pcm

SpeechMarkType

data SpeechMarkType

Constructors

Sentence

Ssml

Viseme

Word

TextType

data TextType

Constructors

TTSsml

TTText

VoiceId

data VoiceId

Constructors

Amy

Astrid

Brian

Carla

Carmen

Celine

Chantal

Conchita

Cristiano

Dora

Emma

Enrique

Ewa

Filiz

Geraint

Giorgio

Gwyneth

Hans

Ines

Ivy

Jacek

Jan

Joanna

Joey

Justin

Karl

Kendra

Kimberly

Liv

Lotte

Mads

Maja

Marlene

Mathieu

Maxim

Miguel

Mizuki

Naja

Nicole

Penelope

Raveena

Ricardo

Ruben

Russell

Salli

Tatyana

Vicki

Vitoria

Lexicon

data Lexicon

Provides lexicon name and lexicon content in string format. For more information, see Pronunciation Lexicon Specification (PLS) Version 1.0 .

See: lexicon smart constructor.

lexicon :: Lexicon

Creates a value of Lexicon with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lContent - Lexicon content in string format. The content of a lexicon must be in PLS format.

lName - Name of the lexicon.

lContent :: Lens' Lexicon (Maybe Text)

Lexicon content in string format. The content of a lexicon must be in PLS format.

lName :: Lens' Lexicon (Maybe Text)

Name of the lexicon.

LexiconAttributes

data LexiconAttributes

Contains metadata describing the lexicon such as the number of lexemes, language code, and so on. For more information, see Managing Lexicons .

See: lexiconAttributes smart constructor.

lexiconAttributes :: LexiconAttributes

Creates a value of LexiconAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

laLanguageCode - Language code that the lexicon applies to. A lexicon with a language code such as "en" would be applied to all English languages (en-GB, en-US, en-AUS, en-WLS, and so on.

laSize - Total size of the lexicon, in characters.

laLexemesCount - Number of lexemes in the lexicon.

laLexiconARN - Amazon Resource Name (ARN) of the lexicon.

laAlphabet - Phonetic alphabet used in the lexicon. Valid values are ipa and x-sampa .

laLastModified - Date lexicon was last modified (a timestamp value).

laLanguageCode :: Lens' LexiconAttributes (Maybe LanguageCode)

Language code that the lexicon applies to. A lexicon with a language code such as "en" would be applied to all English languages (en-GB, en-US, en-AUS, en-WLS, and so on.

laSize :: Lens' LexiconAttributes (Maybe Int)

Total size of the lexicon, in characters.

laLexemesCount :: Lens' LexiconAttributes (Maybe Int)

Number of lexemes in the lexicon.

laLexiconARN :: Lens' LexiconAttributes (Maybe Text)

Amazon Resource Name (ARN) of the lexicon.

laAlphabet :: Lens' LexiconAttributes (Maybe Text)

Phonetic alphabet used in the lexicon. Valid values are ipa and x-sampa .

laLastModified :: Lens' LexiconAttributes (Maybe UTCTime)

Date lexicon was last modified (a timestamp value).

LexiconDescription

data LexiconDescription

Describes the content of the lexicon.

See: lexiconDescription smart constructor.

lexiconDescription :: LexiconDescription

Creates a value of LexiconDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldAttributes - Provides lexicon metadata.

ldName - Name of the lexicon.

ldAttributes :: Lens' LexiconDescription (Maybe LexiconAttributes)

Provides lexicon metadata.

ldName :: Lens' LexiconDescription (Maybe Text)

Name of the lexicon.

Voice

data Voice

Description of the voice.

See: voice smart constructor.

voice :: Voice

Creates a value of Voice with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vLanguageCode - Language code of the voice.

vLanguageName - Human readable name of the language in English.

vGender - Gender of the voice.

vName - Name of the voice (for example, Salli, Kendra, etc.). This provides a human readable voice name that you might display in your application.

vId - Amazon Polly assigned voice ID. This is the ID that you specify when calling the SynthesizeSpeech operation.

vLanguageCode :: Lens' Voice (Maybe LanguageCode)

Language code of the voice.

vLanguageName :: Lens' Voice (Maybe Text)

Human readable name of the language in English.

vGender :: Lens' Voice (Maybe Gender)

Gender of the voice.

vName :: Lens' Voice (Maybe Text)

Name of the voice (for example, Salli, Kendra, etc.). This provides a human readable voice name that you might display in your application.

vId :: Lens' Voice (Maybe VoiceId)

Amazon Polly assigned voice ID. This is the ID that you specify when calling the SynthesizeSpeech operation.